addAdminPage

+++ src/main/java/com/matrix/shopping/dao/repository/ContactRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContactRepository extends CrudRepository<ContactEntity, Long> {
+public interface ContactRepository extends CrudRepository<ContactEntity, Integer> {
 }
