+++ src/main/java/com/matrix/shopping/dao/repository/login/UserRepository.java
@@ -3,7 +3,7 @@ package com.matrix.shopping.dao.repository.login;
 import com.matrix.shopping.dao.entity.login.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
     UserEntity findByUsername(String username);
 
 }
